CTU: Use the float instead of int, which would always be zero when under one.

+++ discord_tron_client/classes/image_manipulation/pipeline.py
@@ -400,7 +400,7 @@ class PipelineRunner:
         return new_image
     
     def _controlnet_all_images(self, preprocessed_images: list, user_config: dict, generator):
-        if int(user_config['strength']) == 0:
+        if float(user_config['strength']) == 0.0:
             # Zero strength = Zero CTU.
             return preprocessed_images
